requirements.txt
streamlit==1.38.0
ultralytics>=8.0.0
Pillow>=10.0.0
numpy>=1.24.0
PyYAML>=6.0
